🚧 Request parsing: functional parsing, seems to work fine for now with several client requests. URI fragments seem to be useless (can keep it for now but considering user-agents don't send this part or the URI when     making a request, will most likely have to remove it).     Next (1): More testing (web browsers, netcat, our cpp client)     Next (2): Parse hexadecimals in uri     Next (3): Create response setter

+++ srcs/http/client_related/client/Client.cpp
@@ -19,7 +19,7 @@ Client::~Client(void)
     delete _address;
 
     size_t queue_size = _requests->size();
-    for (int i = 0; i < queue_size; i++)
+    for (size_t i = 0; i < queue_size; i++)
     {
         delete _requests->front();
     }
